requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==46.0.3
deprecation==2.1.0
dill==0.3.8
einops==0.8.0
exceptiongroup==1.3.0
fastapi==0.120.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.27.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
joblib==1.3.2
kornia==0.7.4
kornia_rs==0.1.8
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pandas==2.2.0
pillow==11.1.0
postgrest==2.24.0
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.11
py4j==0.10.9.7
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pypmml==0.9.17
python-dateutil==2.8.2
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
realtime==2.24.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.4.0
scipy==1.12.0
sentencepiece==0.2.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sklearn-pandas==2.2.0
sklearn2pmml==0.102.0
sniffio==1.3.1
soundfile==0.12.1
spandrel==0.4.0
starlette==0.49.1
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.13.1
threadpoolctl==3.2.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchsde==0.2.6
torchvision==0.20.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.47.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==2.3.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
